refactored all github code into github.js (closes #52)
